[refector] HomeBoard -> CommunityBoard로 뷰모델 이동(북마크 관련 데이터베이스 저장 관련)
